+++ src/main/scala/services/backfill/BackfillOverwriteBatchFactory.scala
@@ -1,5 +1,5 @@
 package com.sneaksanddata.arcane.framework
-package services.streaming.base
+package services.backfill
 
 import models.batches.StagedBackfillOverwriteBatch
 import models.settings.TableName
