+++ src/main/scala/services/backfill/BackfillStreamingOverwriteDataProvider.scala
@@ -1,15 +1,19 @@
 package com.sneaksanddata.arcane.framework
-package services.streaming.base
+package services.backfill
+
+import models.batches.StagedBackfillOverwriteBatch
 
 import zio.Task
 
 /** A trait that represents a backfill data provider.
   */
-trait BackfillStreamingMergeDataProvider:
+trait BackfillStreamingOverwriteDataProvider:
+
+  type BatchType = StagedBackfillOverwriteBatch | Unit
 
   /** Provides the backfill data.
     *
     * @return
     *   A task that represents the backfill data.
     */
-  def requestBackfill: Task[Unit]
+  def requestBackfill: Task[BatchType]
