+++ src/main/scala/services/backfill/GenericBackfillStreamingMergeDataProvider.scala
@@ -1,10 +1,10 @@
 package com.sneaksanddata.arcane.framework
-package services.streaming.base
+package services.backfill
 
 import logging.ZIOLogAnnotations.zlog
 import services.app.base.StreamLifetimeService
+import services.backfill.{BackfillOverwriteBatchFactory, BackfillStreamingMergeDataProvider, BackfillSubStream}
 import services.metrics.base.MetricTagProvider
-import services.streaming.base.{BackfillOverwriteBatchFactory, BackfillStreamingMergeDataProvider, BackfillSubStream}
 
 import zio.stream.ZPipeline
 import zio.{Task, ZIO, ZIOAspect, ZLayer}
