+++ src/main/scala/services/backfill/GenericBackfillStreamingOverwriteDataProvider.scala
@@ -1,5 +1,5 @@
 package com.sneaksanddata.arcane.framework
-package services.streaming.base
+package services.backfill
 
 import logging.ZIOLogAnnotations.zlog
 import models.app.PluginStreamContext
@@ -7,6 +7,7 @@ import models.batches.{MergeableBatch, StagedVersionedBatch}
 import models.schemas.ArcaneSchema
 import models.settings.staging.StagingTableSettings
 import services.app.base.StreamLifetimeService
+import services.backfill.{BackfillOverwriteBatchFactory, BackfillStreamingOverwriteDataProvider, BackfillSubStream}
 import services.metrics.base.MetricTagProvider
 import services.streaming.base.*
 import services.streaming.processors.transformers.StagingProcessor
