+++ src/test/scala/tests/services/streaming/data_providers/backfill/GenericBackfillStreamingOverwriteDataProviderTests.scala
@@ -10,12 +10,7 @@ import services.filters.FieldsFilteringService
 import services.iceberg.{IcebergEntityManager, IcebergS3CatalogWriter, IcebergTablePropertyManager}
 import services.metrics.base.MetricTagProvider
 import services.metrics.{DeclaredMetrics, GlobalMetricTagProvider}
-import services.streaming.base.{
-  BackfillOverwriteBatchFactory,
-  BackfillStreamingOverwriteDataProvider,
-  GenericBackfillStreamingOverwriteDataProvider,
-  StreamDataProvider
-}
+import services.streaming.base.StreamDataProvider
 import services.streaming.graph_builders.GenericStreamingGraphBuilder
 import services.streaming.processors.batch_processors.maintenance.TargetMaintenanceProcessor
 import services.streaming.processors.batch_processors.streaming.{
@@ -27,6 +22,7 @@ import services.streaming.processors.batch_processors.streaming.{
 import services.streaming.processors.transformers.{FieldFilteringTransformer, StagingProcessor}
 import tests.services.streaming.processors.utils.TestStageVersionedBatch
 import tests.shared.*
+import com.sneaksanddata.arcane.framework.services.backfill.{BackfillOverwriteBatchFactory, BackfillStreamingOverwriteDataProvider, GenericBackfillStreamingOverwriteDataProvider}
 
 import org.easymock.{Capture, EasyMock}
 import org.easymock.EasyMock.{replay, verify}
