+++ src/test/scala/tests/services/streaming/graph_builders/GenericGraphBuilderFactoryTests.scala
@@ -5,26 +5,14 @@ import models.settings.backfill.BackfillBehavior
 import services.app.base.StreamLifetimeService
 import services.streaming.base.*
 import services.streaming.graph_builders.GenericGraphBuilderFactory
-import services.streaming.processors.batch_processors.backfill.{
-  BackfillApplyBatchProcessor,
-  BackfillOverwriteWatermarkProcessor
-}
-import services.streaming.processors.batch_processors.streaming.{
-  DisposeBatchProcessor,
-  MergeBatchProcessor,
-  SchemaMigrationProcessor,
-  WatermarkProcessor
-}
+import services.streaming.processors.batch_processors.backfill.{BackfillApplyBatchProcessor, BackfillOverwriteWatermarkProcessor}
+import services.streaming.processors.batch_processors.streaming.{DisposeBatchProcessor, MergeBatchProcessor, SchemaMigrationProcessor, WatermarkProcessor}
 import services.streaming.processors.transformers.{FieldFilteringTransformer, StagingProcessor}
-import tests.shared.{
-  CustomTestBackfillTableSettings,
-  TestPluginBackfillMergeStreamContext,
-  TestPluginBackfillOverwriteStreamContext,
-  TestPluginStreamContext
-}
+import tests.shared.{CustomTestBackfillTableSettings, TestPluginBackfillMergeStreamContext, TestPluginBackfillOverwriteStreamContext, TestPluginStreamContext}
 import services.streaming.batching.StagedBatchFactory
-
 import services.streaming.processors.batch_processors.maintenance.TargetMaintenanceProcessor
+
+import com.sneaksanddata.arcane.framework.services.backfill.{BackfillStreamingMergeDataProvider, BackfillStreamingOverwriteDataProvider}
 import org.scalatest.flatspec.AsyncFlatSpec
 import org.scalatest.matchers.must.Matchers
 import org.scalatest.prop.TableDrivenPropertyChecks.forAll
